+++ packages/swr/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.1](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.4.0...@wundergraph/swr@0.4.1) (2022-10-18)
+
+**Note:** Version bump only for package @wundergraph/swr
+
 ## [0.4.0](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.3.10...@wundergraph/swr@0.4.0) (2022-10-18)
 
 ### Features